requirements.txt
flask
google-generativeai
flask-cors
python-dotenv
google-generativeai